refactor(ci): GitHub CI workflow name 🦖
The name of the GitHub Continuous Integration workflow has been simplified from "Continuous Integration Build 🏗️ and Test 🧪" to "CI Build". This change was made to maintain consistency and simplicity across all GitHub workflows. In addition, the use of emojis and explicit naming of build and test processes were removed for better readability and easier workflow identification.