refactor(ci): remove pull request checkout, change how badges are committed
The changes remove the pull request checkout in workflow since it was unnecessary and could lead to some confusions. Also replaced the old method of committing badges using "EndBug/add-and-commit" action with a more descriptive and clear manual Git commit. This commit provides a clearer commit message and will ease further debugging. It enhances readability and performance of our CI pipeline.