CI: Revert to use EndBug/add-and-commit action for coverage badge update (#9)
* chore(deps): update redis docker tag to v6.2.12

* refactor(ci): Revert to use EndBug/add-and-commit action for coverage badge update

The section of the continuous-integration workflow responsible for committing and pushing any changes to the JaCoCo coverage badge has been modified. Instead of using a custom git command sequence to process changes, this updates the workflow definition to use the EndBug/add-and-commit action, which should offload complexity. This provides a cleaner approach and relies on an established third-party action for this common task.

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>