chore(deps): Updated dependencies and their references in build files
The changes include cleaning up the dependencies in `mindsync.gradle.kts` by replacing concrete versions of dependencies with labels from the dependency catalog, and including the respective aliases from `libs.versions.toml`.

Also, uncommented the previously commented out line to include lib.plugin.buildtools.native in the list of plugins in `mindsync.gradle.kts`.

Updated the file paths in `mindsync-frontend.gradle.kts` to be more consistent and readable.

Added `kotlin-reflect` to the list of dependencies in the `kotlin-conventions.gradle.kts` and `spring-conventions.gradle.kts`.

In `spring-conventions.gradle.kts`, enabled Spring Boot dependency management and added related dependencies.

The aim of these changes is to enhance the ease of managing dependencies and to ensure the build files are tidy and easy to navigate. The update would contribute to simplified maintenance and upgrades of the project in the future.