Create Frontend Template (#22)
* docs(readme): Add CI Build status badge to README

A Continuous Integration (CI) Build status badge was added to the README.md file. This badge will provide visibility of the build status of the project, enhancing transparency for developers and stakeholders.

Closes: 8

* fix(deps): update dependency org.jacoco:org.jacoco.core to v0.8.10

* fix(deps): update dependency org.keycloak:keycloak-admin-client to v21.1.2

* fix(deps): update dependency io.projectreactor:reactor-test to v3.5.8

* fix(deps): update rest.assured to v5.3.1

* fix(deps): update spring boot to v3.1.1

* chore(deps): update dependency gradle to v8.2.1

* fix(deps): update dependency com.github.dasniko:testcontainers-keycloak to v2.6.0

* fix(deps): update dependency com.squareup.moshi:moshi to v1.15.0

* fix(deps): update dependency io.cucumber:cucumber-java to v7.13.0

* chore(deps): update jamesives/github-pages-deploy-action action to v4.4.3

* feat(frontend): Add new components, update configuration files, and improve SEO.

Addition of new components (Menu, HamMenu, HeadSEO) and models (Menu model), changes in existing configuration, and enhancement in the page's SEO.

Newly added components are meant to enhance user interaction and provide a better layout. Changes in configuration files were necessary for introducing these components and for other build specifications. Added SEO friendly tags in the newly created HeadSEO component to improve the web visibility and search-ability.

Add new frontend components and update configurations

Added Menu, HamMenu, and HeadSEO components along with Menu model to enhance user interaction, layout, and website SEO. Made changes to the configuration files such as build.gradle.kts, WebConfig.kt and SecurityConfiguration.kt for the new components and updated build specifications. SEO friendly tags were added in HeadSEO to improve web visibility and search-ability.

* frontend template

* fix(frontend): Sonarqube improvements

* refactor(ci): 🚀 Update commit parameters in continuous-integration.yml

This commit revises the parameters for the commit command in the GitHub Actions script. The change is designed to improve the system's error handling and strengthen security on the project. Instead of adding SVG files, the updated version fetches tags forcefully, ignores pathspec errors, enables push, and uses a secret GitHub token for authentication. The secret token is stored securely in GitHub and accessible only to the GitHub Actions platform. The changes ensure a more robust and secure setup for continuous integration workflows.

* refactor(ci): 🚀 Update checkout action in CI workflow

The checkout action in the continuous integration workflow has been updated to checkout the head ref of the pull request. This change is crucial because it ensures the CI runs against the most updated commit of the PR, rather than the default branch. This way, we can verify changes more efficiently before they are merged.

* commit badge 🍀 %

* refactor(ci): 🚀 Update commit badge message formatting

The commit message formatting in the continuous-integration.yml file has been updated. A space was added between the output placeholder and the percentage sign for better clarity and readability.

* refactor(ci): 🚀 Update commit badge message formatting

The commit message formatting in the continuous-integration.yml file has been updated. A space was added between the output placeholder and the percentage sign for better clarity and readability.

* chore(libs): Update rest-assured version

Updated the version of rest-assured library from 5.3.0 to 5.3.1 in the libs.versions.toml. This update is to get the latest bug fixes and feature enhancements for rest-assured.

* chore(deps): Updated dependencies and their references in build files

The changes include cleaning up the dependencies in `mindsync.gradle.kts` by replacing concrete versions of dependencies with labels from the dependency catalog, and including the respective aliases from `libs.versions.toml`.

Also, uncommented the previously commented out line to include lib.plugin.buildtools.native in the list of plugins in `mindsync.gradle.kts`.

Updated the file paths in `mindsync-frontend.gradle.kts` to be more consistent and readable.

Added `kotlin-reflect` to the list of dependencies in the `kotlin-conventions.gradle.kts` and `spring-conventions.gradle.kts`.

In `spring-conventions.gradle.kts`, enabled Spring Boot dependency management and added related dependencies.

The aim of these changes is to enhance the ease of managing dependencies and to ensure the build files are tidy and easy to navigate. The update would contribute to simplified maintenance and upgrades of the project in the future.

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>