feat(ci): Update CI workflow and Gradle properties
The continuous integration (.github/workflows/continuous-integration.yml) workflow has been updated to use the latest version of actions/setup-node version 3 and to automatically fetch the node's long-term support (LTS) version. An upgrade of the actions/setup-java to version 3 and an update in the working JDK to 17 has also been implemented. Icons for 'Update Version in gradle.properties' and 'Publish Docs' were added to improve readability and understandability. The Docker build and push action has been updated for continuous integration and deployment (CI/CD).

 In the Gradle file (build.gradle.kts), the "group" properties were dynamically assigned to facilitate more flexible project configurations.