refactor(ci): Improve readability and workflow in Continuous Integration config
This update improves the overall readability of the continuous integration configuration file. In addition to updating some of the labels with more informative and visually apparent text, some other changes include:

- Changed OS specification for compilation from Ubuntu 20.04 to the latest version of Ubuntu.
- Remove the use of 'actions' in favour of direct command runs for better usability and comprehension of what each step is accomplishing.
- Used a better method to run Gradle tasks that integrated the steps into one command, making it cleaner and easier to understand.
- With security checks, instead of using a generic term, made it clear that it's specifically OWASP Dependency Check.
- Finally, ordered the name of jobs and steps in a way that describes the purpose better.

These changes allow for more efficient maintenance and comprehension, particularly for newcomers or external contributors. This update does not affect the CI's functional properties or processes.