test(auth): Add new test case and update workflow configuration ⚙️
Changes were made in the KeycloakRepositoryTest for better user creation failure handling. A new test case was added to examine behaviour when Keycloak returns an error during user creation. This allows us to handle errors more efficiently and improve system resilience.

Additionally, an 'if' condition was added to the ci configuration file to ensure that the commit and push action for coverage badge only runs on the 'master' branch. This prevents unnecessary commits and ensures that the badge accurately represents the status of the main branch.