docs(core): Implement ApiDataResponse and update swagger documentation
The ApiResponse was renamed to ApiDataResponse for a better explanation of its purpose. ObjectMapperConfig was added to enhance JSON object mappings. Dependencies in mindsync.gradle.kts were updated along with changes in path of Swagger UI for springdoc in application.yml. Extended swagger documentation was implemented for individual API endpoints for a clear understanding of their functionality and expected response codes. Changes in UserRegisterController and UserRegistrator to accommodate ApiResponse to ApiDataResponse.

These changes will provide better readability and understanding of the application's APIs for the developers and users, assisting in easier tracking and user experience. It also helps in better categorization of request and response objects, further enhancing code organization.