fix(auth): 🚨 Update UserAuthenticatorController for reactive programming
Implemented a change to the login function within the UserAuthenticatorController using the Reactive Streams' standard: Mono. This is part of updating the application to utilize a more efficient, non-blocking back-pressure algorithm. Also, made use of more explicit constant for the login route. This clarification provides cleaner and easier to understand code.