[TASK-16] docs(core): Implement ApiDataResponse and update swagger documentation (#94)
* docs(core): Implement ApiDataResponse and update swagger documentation

The ApiResponse was renamed to ApiDataResponse for a better explanation of its purpose. ObjectMapperConfig was added to enhance JSON object mappings. Dependencies in mindsync.gradle.kts were updated along with changes in path of Swagger UI for springdoc in application.yml. Extended swagger documentation was implemented for individual API endpoints for a clear understanding of their functionality and expected response codes. Changes in UserRegisterController and UserRegistrator to accommodate ApiResponse to ApiDataResponse.

These changes will provide better readability and understanding of the application's APIs for the developers and users, assisting in easier tracking and user experience. It also helps in better categorization of request and response objects, further enhancing code organization.

* feat(core): ObjectMapperConfig for consistent property naming 🐍

Added PropertyNamingStrategy to ObjectMapperConfig in mindsync app to adopt consistent SNAKE_CASE property naming across the API JSON responses. This change ensures that API consumers have a predictable data structure and promotes easier data processing on the consumer side.

* feat(core): 🚨 Add more detailed exception handling and improve security configs

This commit includes multiple enhancements to the code. The GlobalExceptionHandler class has been updated, adding handlers for IllegalArgumentException and generic Exception types, providing more detailed information about errors when they occur. A minor optimization was also made in the Role class, replacing the redundant values() call with entries. The order of response codes in UserAuthenticatorController was also tidied up.

On the security side, CSRF protection was enabled in the application.yml to guard against specific types of attacks. Role-based access is now configured for select paths in the security configuration file, enhancing the security of the application.

These changes will help in better error reporting and improving the security layer of the application.

* fix(auth): 🚨 Update UserAuthenticatorController for reactive programming

Implemented a change to the login function within the UserAuthenticatorController using the Reactive Streams' standard: Mono. This is part of updating the application to utilize a more efficient, non-blocking back-pressure algorithm. Also, made use of more explicit constant for the login route. This clarification provides cleaner and easier to understand code.