docs(swagger) Update Springdoc dependencies and adjust code accordingly
Updated Springdoc dependencies versions to 2.2.0 in the MindSync project due to a necessary upgrade. The changes in the configuration file and the SpringdocConfiguration class were necessary due to the updated versions. The content security policy in the ApplicationSecurityPropertiesTest has been slightly modified to match the new Springdoc versions and to increase security efficiency.