refactor(apps/mindsync-frontend): ² Add additional script commands to package.json
Scripts for checking the landing page and running linters on the frontend have been added for ease of use. This was done to keep all commands in one reliable place, making it easier for the development team to run tests and ensure the code consistently meets the project's standards. Additionally reduced repetition in html file.