chore(deps): ◌ Bump up versions of springBoot and testcontainers
This commit increases the version of springBoot from 3.1.2 to 3.1.3 and testcontainers from 1.18.3 to 1.19.0 in gradle/libs.versions.toml. This update is necessary to ensure compatibility with latest updates and to benefit from bug fixes and new features of these libraries. Please make sure to test the application thoroughly after this update.