test(apps/client): 🧪 Remove deprecated test setup file, improve form validation, rename files, refactor tests
Deleted the HelloWorld.example.ts file as it was a placeholder file and wasn't necessary for the test setup.

Introduced Vee-Validate package for form validation in the LoginView component. This has resulted in much cleaner code and more precise validation rules for each input field in the form.

Renamed `MockResponse.ts` to `ResponseMocks.ts` and `MockResponse.ts` to `ResponseMocks.ts` for better clarity.

Refactored the test setup files to be more intelligible and maintainable. Introduced new helper files like `AccessTokenMocks.ts` and `UserMocks.ts` which take care of generating mock data that can be used in tests.

Many test cases in the Authentication module were refactored to include the usage of these helper functions, thereby making the tests cleaner and more maintainable.

Added new test cases for LoginView to cover form validation and form submission.

The refactoring and renaming